fix: add /api/v1 prefix to routes
the current gateway does not strip the prefix from the url. adding fastapi root_path will interfere with swagger html.

The only solution is to add this prefix to each of the routers. the best practice is to strip the prefix in the gateway. it might get updated after changing the gateway.